pyproject.toml
[tool.poetry]
name = "votr"
version = "0.1.0"
description = ""
authors = ["Tom Curran <tcurran4589@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiofiles = "23.2.1"
aioredis = "2.0.0"
aiosqlite = "0.17.0"
annotated-types = "0.6.0"
anyio = "4.2.0"
appnope = "0.1.2"
argon2-cffi = "21.3.0"
argon2-cffi-bindings = "21.2.0"
arrow = "1.3.0"
asgi-correlation-id = "4.3.1"
asgiref = "3.5.2"
asttokens = "2.0.5"
async-lru = "2.0.4"
async-timeout = "4.0.3"
asyncpg = "0.29.0"
attrs = "23.1.0"
babel = "2.11.0"
backcall = "0.2.0"
bcrypt = "4.1.3"
beautifulsoup4 = "4.12.2"
bleach = "4.1.0"
brotli = "1.0.9"
build = "1.2.1"
cachecontrol = "0.14.0"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "2.0.4"
cleo = "2.1.0"
click = "8.1.7"
colorlog = "6.8.2"
comm = "0.2.1"
crashtest = "0.4.1"
debugpy = "1.6.7"
decorator = "5.1.1"
defusedxml = "0.7.1"
distlib = "0.3.8"
django = "4.1"
dnspython = "2.6.1"
docopt = "0.6.2"
dulwich = "0.21.7"
email-validator = "2.1.1"
executing = "0.8.3"
fastapi = "0.103.0"
fastapi-admin = "1.0.4"
fastjsonschema = "2.19.1"
filelock = "3.14.0"
fqdn = "1.5.1"
googlemaps = "4.10.0"
h11 = "0.14.0"
hiredis = "2.3.2"
httpcore = "1.0.5"
httplib2 = "0.22.0"
httptools = "0.6.1"
httpx = "0.27.0"
idna = "3.7"
importlib-metadata = "7.1.0"
installer = "0.7.0"
ipykernel = "6.28.0"
ipython = "8.12.3"
ipywidgets = "8.1.2"
iso8601 = "1.1.0"
isoduration = "20.11.0"
itsdangerous = "2.2.0"
jaraco-classes = "3.4.0"
jedi = "0.18.1"
jinja2 = "3.1.3"
json5 = "0.9.6"
jsonpointer = "2.4"
jsonschema = "4.19.2"
jsonschema-specifications = "2023.7.1"
keyring = "24.3.1"
loguru = "0.7.2"
markdown-it-py = "3.0.0"
markupsafe = "2.1.3"
matplotlib-inline = "0.1.6"
mdurl = "0.1.2"
mistune = "2.0.4"
more-itertools = "10.2.0"
msgpack = "1.0.8"
nbclient = "0.8.0"
nbconvert = "7.16.4"
nbformat = "5.9.2"
nest-asyncio = "1.6.0"
notebook = "7.0.8"
notebook-shim = "0.2.3"
opensecrets-crpapi = "0.2.2"
orjson = "3.10.3"
overrides = "7.4.0"
packaging = "23.2"
pandocfilters = "1.5.0"
parso = "0.8.3"
pendulum = "3.0.0"
pexpect = "4.8.0"
pickleshare = "0.7.5"
pkginfo = "1.10.0"
platformdirs = "3.10.0"
ply = "3.11"
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.8.0"
prometheus-client = "0.14.1"
prompt-toolkit = "3.0.43"
psutil = "5.9.0"
psycopg2 = "2.9.9"
psycopg2-binary = "2.9.9"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pycparser = "2.21"
pydantic = "2.5.3"
pydantic-extra-types = "2.7.0"
pydantic-settings = "2.2.1"
pydantic-core = "2.14.6"
pygeocodio = "1.4.0"
pygments = "2.15.1"
pyparsing = "3.1.2"
pypika-tortoise = "0.1.6"
pyproject-hooks = "1.1.0"
pysocks = "1.7.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-json-logger = "2.0.7"
python-multipart = "0.0.9"
pytz = "2024.1"
pyyaml = "6.0.1"
pyzmq = "25.1.2"
qtconsole = "5.5.1"
qtpy = "2.4.1"
rapidfuzz = "3.9.1"
redis = "5.0.4"
referencing = "0.30.2"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
rfc3339-validator = "0.1.4"
rfc3986-validator = "0.1.1"
rich = "13.7.1"
rpds-py = "0.10.6"
send2trash = "1.8.2"
shellingham = "1.5.4"
sip = "6.7.12"
six = "1.16.0"
sniffio = "1.3.0"
soupsieve = "2.5"
sqlalchemy = "2.0.30"
sqlparse = "0.4.4"
stack-data = "0.2.0"
starlette = "0.27.0"
terminado = "0.17.1"
time-machine = "2.14.1"
tinycss2 = "1.2.1"
tomlkit = "0.12.5"
tornado = "6.3.3"
tortoise-orm = "0.21.2"
traitlets = "5.7.1"
trove-classifiers = "2024.5.22"
types-python-dateutil = "2.9.0.20240316"
typing-extensions = "4.11.0"
tzdata = "2024.1"
ujson = "5.10.0"
uri-template = "1.3.0"
urllib3 = "2.2.1"
uvicorn = "0.20.0"
uvloop = "0.19.0"
virtualenv = "20.26.2"
watchfiles = "0.22.0"
wcwidth = "0.2.5"
webcolors = "1.13"
webencodings = "0.5.1"
websocket-client = "1.8.0"
websockets = "12.0"
widgetsnbextension = "4.0.10"
xattr = "1.1.0"
yarg = "0.1.9"
zipp = "3.19.0"
pycongress = "^0.0.1"
google-api-python-client = "^2.132.0"
phonenumbers = "^8.13.38"
cryptography = "^42.0.8"
keycove = "^0.3.5"
gunicorn = "^22.0.0"
slowapi = "^0.1.9"
fastapi-limiter = "^0.1.6"
pandas = "^2.2.2"
polars = "^0.20.31"
numpy = "^2.0.0"
fastapi-cache2 = "^0.2.1"
numba = "^0.60.0"
alembic = "^1.13.2"
greenlet = "^3.0.3"
strawberry-graphql = {extras = ["fastapi"], version = "^0.237.2"}
fastapi-sessions = "^0.3.2"
us = "^3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.alembic]
script_location = "alembic"
file_template = "%(rev)s_%(slug)s"
sqlalchemy.url = "$DEV_DATABASE_URL"  # Replace with your actual connection string

# Logging Configuration (Optional):
[tool.alembic.logging]
level = "INFO"  # Set the desired logging levela
file = "alembic.log"  # Log to a file (optional)

# Recommended Alembic Configuration Options:
# Comma-separated list of version locations
version_locations = ["alembic/versions"] 

# Exclude tables that should not be managed by Alembic
exclude_tables = ["spatial_ref_sys"]
